Misc minor fixups.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@5973 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/common/common.c
@@ -4282,7 +4282,7 @@ char *COM_ParseStringSet (const char *data, char *out, size_t outsize)
 }
 
 
-char *COM_ParseType (const char *data, char *out, int outlen, com_tokentype_t *toktype)
+char *COM_ParseType (const char *data, char *out, size_t outlen, com_tokentype_t *toktype)
 {
 	int		c;
 	int		len;
